Moving create_dispatch_mem_map to qa_hal_dispatch.cpp
