+++ tt_metal/llrt/hal/tt-2xx/sources.cmake
@@ -1,6 +1,7 @@
 set(QA_HAL_SOURCES
     quasar/qa_hal.cpp
     quasar/qa_hal_tensix.cpp
+    quasar/qa_hal_dispatch.cpp
     quasar/qa_hal_active_eth.cpp
     quasar/qa_hal_idle_eth.cpp
 )
